skill/requirements.txt
pyyaml>=6.0
pytz>=2024.1
